+++ internal/provider/st_admin_role_resource_test.go
@@ -1,0 +1,81 @@
+package provider
+
+import (
+	"fmt"
+	"testing"
+	"time"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+)
+
+func TestNewSTAdminRoleResource(t *testing.T) {
+	t.Parallel()
+	r := time.Now().Format("-2006-01-02_15-04-05") + "-" + fmt.Sprint(time.Now().UnixNano())
+	resourceType := "xmft_st_admin_role"
+	resourceName := "role21"
+	name := "role1" + r
+
+	localConfig := ``
+
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// Create and Read testing
+			{
+				Config: providerConfig + localConfig + `
+					resource "` + resourceType + `" "` + resourceName + `" {
+						provider         = xmft.st1
+						name             = "` + name + `"
+						is_limited       = true
+						is_bounce_allowed = true
+						menus             = [ "Server Log" ]
+					}
+					`,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					checkResources([]byte(`
+					resource "`+resourceType+`" "`+resourceName+`" {
+						#provider        = xmft.st1
+						name             = "`+name+`"
+						is_limited       = true
+						is_bounce_allowed = true
+					}
+					`)),
+					resource.TestCheckResourceAttrSet(resourceType+"."+resourceName, "last_updated"),
+				),
+			},
+			/*// ImportState testing
+			{
+				ResourceName:      "xmft_staccount.account1",
+				ImportState:       true,
+				ImportStateVerify: true,
+				// The last_updated attribute does not exist in the HashiCups
+				// API, therefore there is no value for it during import.
+				ImportStateVerifyIgnore: []string{"last_updated"},
+			},*/
+			// Update and Read testing
+			{
+				Config: providerConfig + localConfig + `
+				resource "` + resourceType + `" "` + resourceName + `" {
+					    provider         = xmft.st1
+						name             = "` + name + `"
+						is_limited       = false
+						is_bounce_allowed = true
+						menus             = [ "Server Log", "File Tracking"	 ]
+				}
+				`,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					checkResources([]byte(`
+					resource "`+resourceType+`" "`+resourceName+`" {
+						#provider        = xmft.st1
+						name             = "`+name+`"
+						is_limited       = false
+						is_bounce_allowed = true
+					}
+					`)),
+					resource.TestCheckResourceAttrSet(resourceType+"."+resourceName, "last_updated"),
+				),
+			},
+			// Delete testing automatically occurs in TestCase
+		},
+	})
+}