+++ internal/provider/st_conf_file_archiving_resource_test.go
@@ -1,0 +1,80 @@
+package provider
+
+import (
+	"testing"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+)
+
+func TestNewSTArchivingResource(t *testing.T) {
+	t.Parallel()
+	// r := time.Now().Format("-2006-01-02_15-04-05") + "-" + fmt.Sprint(time.Now().UnixNano())
+	resourceType := "xmft_st_file_archiving"
+	resourceName := "archiving1"
+
+	localConfig := ``
+
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// Create and Read testing
+			{
+				Config: providerConfig + localConfig + `
+					resource "` + resourceType + `" "` + resourceName + `" {
+						provider         = xmft.st1
+						#enabled          = true
+						archive_folder = "/tmp/archive"
+					}
+					`,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					checkResources([]byte(`
+					resource "`+resourceType+`" "`+resourceName+`" {
+						#provider        = xmft.st1
+						#enabled          = true
+						global_archiving_policy = "disabled"
+						archive_folder = "/tmp/archive"
+						delete_files_older_than = 1
+						delete_files_older_than_unit = "days"
+						maximum_file_size_allowed_to_archive = 0
+					}
+					`)),
+					resource.TestCheckResourceAttrSet(resourceType+"."+resourceName, "last_updated"),
+				),
+			},
+			/*// ImportState testing
+			{
+				ResourceName:      "xmft_staccount.account1",
+				ImportState:       true,
+				ImportStateVerify: true,
+				// The last_updated attribute does not exist in the HashiCups
+				// API, therefore there is no value for it during import.
+				ImportStateVerifyIgnore: []string{"last_updated"},
+			},*/
+			// Update and Read testing
+			{
+				Config: providerConfig + localConfig + `
+				resource "` + resourceType + `" "` + resourceName + `" {
+					    provider         = xmft.st1
+						#enabled          = true
+						archive_folder = "/tmp/archive4"
+				}
+				`,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					checkResources([]byte(`
+					resource "`+resourceType+`" "`+resourceName+`" {
+						#provider        = xmft.st1
+						#enabled          = true
+						global_archiving_policy = "disabled"
+						archive_folder = "/tmp/archive4"
+						delete_files_older_than = 1
+						delete_files_older_than_unit = "days"
+						maximum_file_size_allowed_to_archive = 0
+					}
+					`)),
+					resource.TestCheckResourceAttrSet(resourceType+"."+resourceName, "last_updated"),
+				),
+			},
+			// Delete testing automatically occurs in TestCase
+		},
+	})
+}