+++ internal/provider/st_user_class_resource_test.go
@@ -1,0 +1,80 @@
+package provider
+
+import (
+	"fmt"
+	"testing"
+	"time"
+
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+)
+
+func TestSTUserClassResource(t *testing.T) {
+	t.Parallel()
+	r := time.Now().Format("-2006-01-02_15-04-05") + "-" + fmt.Sprint(time.Now().UnixNano())
+	resourceType := "xmft_st_user_class"
+	resourceName := "class1"
+	name := "class1" + r
+
+	localConfig := ``
+
+	resource.Test(t, resource.TestCase{
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps: []resource.TestStep{
+			// Create and Read testing
+			{
+				Config: providerConfig + localConfig + `
+					resource "` + resourceType + `" "` + resourceName + `" {
+						provider         = xmft.st1
+						name             = "` + name + `"
+					}
+					`,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					checkResources([]byte(`
+					resource "`+resourceType+`" "`+resourceName+`" {
+						#provider        = xmft.st1
+						name             = "`+name+`"
+					}
+					`)),
+					resource.TestCheckResourceAttrSet(resourceType+"."+resourceName, "last_updated"),
+				),
+			},
+			/*// ImportState testing
+			{
+				ResourceName:      "xmft_staccount.account1",
+				ImportState:       true,
+				ImportStateVerify: true,
+				// The last_updated attribute does not exist in the HashiCups
+				// API, therefore there is no value for it during import.
+				ImportStateVerifyIgnore: []string{"last_updated"},
+			},*/
+			// Update and Read testing
+			{
+				Config: providerConfig + localConfig + `
+				resource "` + resourceType + `" "` + resourceName + `" {
+					    provider         = xmft.st1
+						name             = "` + name + `"
+						user_type 	     = "real"
+						user_name        = "user1"
+						group            = "group1"
+						address          = "192.168.1.1"
+
+				}
+				`,
+				Check: resource.ComposeAggregateTestCheckFunc(
+					checkResources([]byte(`
+					resource "`+resourceType+`" "`+resourceName+`" {
+						#provider        = xmft.st1
+						name             = "`+name+`"
+						user_type 	     = "real"
+						user_name        = "user1"
+						group            = "group1"
+						address          = "192.168.1.1"
+					}
+					`)),
+					resource.TestCheckResourceAttrSet(resourceType+"."+resourceName, "last_updated"),
+				),
+			},
+			// Delete testing automatically occurs in TestCase
+		},
+	})
+}